factor out retry, i guess
