+++ test/booth.mjs
@@ -4,25 +4,7 @@ import delay from 'delay';
 import supertest from 'supertest';
 import createUwave from './utils/createUwave.mjs';
 import testSource from './utils/testSource.mjs';
-
-/** Retry the `fn` until it doesn't throw, or until the duration in milliseconds has elapsed. */
-async function retryFor(duration, fn) {
-  const end = Date.now() + duration;
-  let caughtError;
-  while (Date.now() < end) {
-    try {
-      const result = await fn();
-      return result;
-    } catch (err) {
-      caughtError = err;
-    }
-    await delay(10);
-  }
-
-  if (caughtError != null) {
-    throw new Error(`Failed after ${duration}ms`, { cause: caughtError });
-  }
-}
+import { retryFor } from './utils/retry.mjs';
 
 describe('Booth', () => {
   describe('GET /booth', () => {
